requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
pinecone-client
sentence-transformers
google-generativeai
pydantic 